feat: migrate Avatar and CoCTeamList components to TypeScript
Fixes #143
Fixes #144
Fixes #145
Fixes #146
Fixes #147

- Converted Avatar and CoCTeamList components from JavaScript to TypeScript, adding type definitions for props.
- Updated tests for Avatar and CoCTeamList to reflect the new import style.
- Enhanced the index file to export the newly typed components.
- Refactored related imports across the application to ensure consistency with the new TypeScript structure.
