+++ frontend/components/index.js
@@ -1,3 +1,8 @@
 export { default as Card } from './Card';
 export { default as Input } from './Input';
-export * from './Table'; 
+export * from './Table';
+export * from './Avatar';
+export * from './CoCTeamList';
+export * from './UserRegistrationForm';
+export * from './ReportForm';
+export * from './ReportDetailView';
